fix: signature extraction command (#8)
* fix: signature extraction command

Fix a bug in command to extract raw (r, s) values from DER encoded
signature. We need to pad the hex strings for r and s with leading 0s if
needed, so that they are 32 bytes in binary.

* Add self-check for signature verification

Add a self-check in openssl-crosscheck.sh to verify generated signature.

This is also a helpful countermeasure against fault injection attacks.